pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "wheatvol"
version = "0.1.0"
description = "Ukraine shock volatility study for wheat (GARCH/EGARCH/GJR/FIGARCH)"
authors = [{name = "Auteur"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas",
  "numpy",
  "matplotlib",
  "statsmodels",
  "arch",
  "scipy",
  "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest", "flake8", "black", "pre-commit"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
